// webapp/resources/sap/esh/search/ui/sinaNexTS/sina/ComparisonOperator.ts
/*!
 * SAPUI5
 * Copyright (c) 2025 SAP SE or an SAP affiliate company. All rights reserved.
 * 
 */
export enum ComparisonOperator {
    Search = "Search",
    Eq = "Eq",
    Ne = "Ne", // not equal
    Gt = "Gt",
    Lt = "Lt",
    Ge = "Ge",
    Le = "Le",
    Co = "Co", // Contains only
    Bw = "Bw",
    Ew = "Ew", // End with
    ChildOf = "ChildOf",
    DescendantOf = "DescendantOf",
}