// webapp/resources/sap/esh/search/ui/sinaNexTS/providers/abap_odata/ComparisonOperator.d.ts
declare module "sap/esh/search/ui/sinaNexTS/providers/abap_odata/ComparisonOperator" {
    /*!
     * SAPUI5
     * Copyright (c) 2025 SAP SE or an SAP affiliate company. All rights reserved.
     *
     */
    enum ABAPODataComparisonOperator {
        Bt = "BT"
    }
}
//# sourceMappingURL=ComparisonOperator.d.ts.map