// webapp/resources/sap/esh/search/ui/sinaNexTS/sina/LogicalOperator.d.ts
declare module "sap/esh/search/ui/sinaNexTS/sina/LogicalOperator" {
    /*!
     * SAPUI5
     * Copyright (c) 2025 SAP SE or an SAP affiliate company. All rights reserved.
     *
     */
    enum LogicalOperator {
        And = "And",
        Or = "Or",
        Not = "Not",
        Row = "Row"
    }
}
//# sourceMappingURL=LogicalOperator.d.ts.map